--- J_Atulya.py
@@ -1,118 +1,0 @@
-import turtle
-
-screen = turtle.Screen()
-circle_turtle = turtle.Turtle()
-circle_turtle.speed(0)
-
-def draw_pattern(n,m,p):
-    circle_turtle.begin_fill()
-    for i in range(m):  
-        for j in range(4):  
-            circle_turtle.forward(p)
-            circle_turtle.right(90)
-        circle_turtle.right(n)  
-    circle_turtle.end_fill()
-
-circle_turtle.color("orange")
-circle_turtle.fillcolor("orange")
-circle_turtle.penup()
-circle_turtle.goto(0, -308)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(300)
-circle_turtle.end_fill()
-
-circle_turtle.color("darkgreen")
-circle_turtle.fillcolor("darkgreen")
-circle_turtle.penup()
-circle_turtle.goto(0, -300)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(290)
-circle_turtle.end_fill()
-
-circle_turtle.color("yellow")
-circle_turtle.fillcolor("yellow")
-circle_turtle.penup()
-circle_turtle.goto(0, 0)
-circle_turtle.pendown()
-draw_pattern(45, 16, 200)
-
-circle_turtle.color("orange")
-circle_turtle.fillcolor("orange")
-circle_turtle.penup()
-circle_turtle.goto(0, 0)
-circle_turtle.pendown()
-draw_pattern(90, 8,200)
-
-circle_turtle.color("darkgreen")
-circle_turtle.fillcolor("darkgreen")
-circle_turtle.penup()
-circle_turtle.goto(0, -200)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(200)  
-circle_turtle.end_fill()
-
-circle_turtle.color("red")
-circle_turtle.fillcolor("red")
-circle_turtle.penup()
-circle_turtle.goto(0, -195)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(195)  
-circle_turtle.end_fill()
-
-circle_turtle.color("yellow")
-circle_turtle.fillcolor("yellow")
-circle_turtle.penup()
-circle_turtle.goto(0, 0)
-circle_turtle.pendown()
-draw_pattern(40, 16, 100)
-
-circle_turtle.color("darkgreen")
-circle_turtle.fillcolor("darkgreen")
-circle_turtle.penup()
-circle_turtle.goto(75, -10)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(79)  
-circle_turtle.end_fill()
-
-circle_turtle.color("orange")
-circle_turtle.fillcolor("orange")
-circle_turtle.penup()
-circle_turtle.goto(71, -10)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(75)  
-circle_turtle.end_fill()
-
-circle_turtle.color("darkgreen")
-circle_turtle.fillcolor("darkgreen")
-circle_turtle.penup()
-circle_turtle.goto(33,-8)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(38)  
-circle_turtle.end_fill()
-
-circle_turtle.color("yellow")
-circle_turtle.fillcolor("yellow")
-circle_turtle.penup()
-circle_turtle.goto(28, -8)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(33)  
-circle_turtle.end_fill()
-
-circle_turtle.color("red")
-circle_turtle.fillcolor("red")
-circle_turtle.penup()
-circle_turtle.goto(2, -5)
-circle_turtle.pendown()
-circle_turtle.begin_fill()
-circle_turtle.circle(10)  
-circle_turtle.end_fill()
-
-screen.exitonclick()
